bluez5_%.bbappend: device: Remove device after all bearers are disconnected
For a dual-mode remote, both BR/EDR and BLE may be connected,
RemoveDevice should be handled after all bearers are disconnects.
Otherwise, if msg is removed, but not all connection are dropped,
this function returns before *remove is updated, then after all
connections are dropped, but device->disconnects is NULL,
remove_device is not updated. Consequently *remove is not set to
true. Remove device is not performed in adapter_remove_connection.

Upstream-Status: BACKPORT

Change-Id: I927875bc88d475858222e4df0609eaae1fff56c9
Signed-off-by: Cheng Jiang <quic_chejiang@quicinc.com>