ci/base: add the yocto upstream sstate-mirror
As the development speed of the Poky master is quite fast,
we can take advantage of the sstate-cahce mirror that is made available.
This brings advantages especially when building the toolchain,
which we do not modify in any way and so we can use the existing one.

The Yocto Project now allows you to reuse the Shared State cache from its autobuilder [1].
If the network connection between our server and your machine is faster than you would build recipes from source,
[1] https://docs.yoctoproject.org/dev/migration-guides/release-notes-4.0.html?highlight=sstate_mirrors#new-features-enhancements-in-4-0

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>