ci: build-yocto: merge the check-layer job
Merge this two jobs since since there are dependencies between them.
They both want to use the same references in the repositories they use
and need the same lock step.

This will also make it possible to add interdependencies if necessary
so that there are jobs A that depend on others B.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>