qcom-partition-confs: add recipe based on qclinux 1.2
Recipe based on kirkstone rev 3ac146c (qclinux 1.2), used to generate the
machine specific partition.xml file which is later converted to binary with
help from ptool.py.

Partition confs for qcm6490 and qcs9100 modified to remove persist and
overlay, and instead use a larger system/rootfs partition, since that
should really be a distro decision and not a BSP one.

Signed-off-by: Ricardo Salveti <ricardo.salveti@oss.qualcomm.com>