linux-yocto-dev: update QCM6490-IDP UFS patch
The patch in this layer that enables UFS nodes on the QCM6490-IDP board
corresponds to the v2 of the patchset which was rejected upstream
(instead of being Pending, as it was incorrectly stated in the
Upstream-Status tag).

Replace it with the version 3 of the patch, which is has been accepted
by Bjorn as is pending for the 6.13 release.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>