ci: yocto-check-layer: ensure KAS workspace is outside of the repo
In the Github actions for yocto-check-layer, we can see the following
error:

INFO: Traceback (most recent call last):
  File "/opt/actions-runner/_work/meta-qcom-hwe/meta-qcom-hwe/poky/scripts/lib/checklayer/cases/common.py", line 87, in test_patches_upstream_status
    self.assertEqual(len(patches), 0 , \
AssertionError: 1 != 0 : Found following patches with malformed or missing upstream status:
/opt/actions-runner/_work/meta-qcom-hwe/meta-qcom-hwe/poky/meta/lib/oeqa/manual/abat.patch

Since KAS_WORK_DIR is not set, and since kas is executed from inside
the repo, we end up cloning poky and meta-qcom inside meta-qcom-hwe
repo. As such yocto-check-layer will 'parse' the entire poky (and
meta-qcom) repo , especially when looking for .patch files. Arguably
yocto-check-layer could be smarter and skip .patch files not
referenced in a valid recipe.. but it's not the case.

So we need to ensure that we checkout the kas workspace outside of our
repo. With this patch we either let the caller manage KAS_WORK_DIR, or
we create a temporary directory ourselves (typically in Github Actions).

This patch fixes the error mentioned above.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>