firmware: set the download filename
The boot firmware files published by Qualcomm, even if they are
versioned properly, are pusblished as
<platform>_bootbinaries.zip. Bitbake expects the version to be there
in the downloaded filenames, since all downloads are placed in the
same location (DL_DIR). As such installing different versions of the
boot binaries will override them each time, and bitbake complains
about that (and it causes unnecessary warnings and re-fetch).

We can customize the download file name using the 'downloadfilename'
attribute in SRC_URI. With this patch, the firmware is downloaded as:

downloads/QCM6490_bootbinaries_r1.0_00005.0.zip

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>