layer.conf: increase BBFILE_PRIORITY to 6
This gets us slightly more deterministic behaviour since most layers,
inclusing OE-core, set it to 5. When priorities are equal, the ordering
in BBLAYERS is used to resolve conflicts. Since this layer was at '5',
getting the ordering wrong would negate .bbappends and shadowed recipes.

Upstream bitbake wants to phase this out, but for the time being, let's
decrease the opportunities for users to get this wrong.

Fixes: https://github.com/quic-yocto/meta-qcom-hwe/issues/98

Signed-off-by: Koen Kooi <koen.kooi@oss.qualcomm.com>