qcom-partition-confs: add support for CDT
For some platforms, the CDT file is published and can be used when
flashing a device with QDL/firehose. Unfortunately, the CDT is not
always available, so we have to twist the recipe and the partition
conf file only on selected platform.

The RB3 Gen2 is the first platform that makes use of the public CDT,
the idea is to fixup the partition.conf file and insert
"--filename=cdt.bin" on the right line.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>